// src/status-paciente/status-paciente.interface.ts
export enum STATUS_PACIENT_ID {
  queue_avaliation = 1,
  queue_therapy = 2,
  therapy = 3,
  avaliation = 4,
  crud_therapy = 5,
}

export enum STATUS_PACIENT_COD {
  queue_avaliation = 'queue_avaliation',
  avaliation = 'avaliation',
  queue_devolutiva = 'queue_devolutiva',
  devolutiva = 'devolutiva',
  queue_therapy = 'queue_therapy',
  therapy = 'therapy',
  crud_therapy = 'crud_therapy',
}